+++ src-input/duk_api_bytecode.c
@@ -487,7 +487,7 @@ static duk_uint8_t *duk__load_func(duk_context *ctx, duk_uint8_t *p, duk_uint8_t
 			duk_double_t val;
 			DUK__ASSERT_LEFT(8);
 			val = DUK_RAW_READ_DOUBLE_BE(p);
-			DUK_TVAL_SET_NUMBER_CHKFAST(&tv_tmp, val);
+			DUK_TVAL_SET_NUMBER_CHKFAST_SLOW(&tv_tmp, val);
 			duk_push_tval(ctx, &tv_tmp);
 			break;
 		}